Moving files pt 1
Moved Register into its own folder in containers/auth.
Moved login into container/auth.
maved App.js and App.css into src root.
