+++ front_end/src/App.js
@@ -1,10 +1,10 @@
 import React, {Component} from 'react';
 import {Route, Switch} from 'react-router-dom';
 
-import Dashboard from './Dashboard/Dashboard';
+import Dashboard from './components/Dashboard/Dashboard';
 import Login from '../components/Login';
-import PrivateRoute from '../common/PrivateRoute';
-import NavBar from './Navigation/NavBar/NavBar';
+import PrivateRoute from './common/PrivateRoute';
+import NavBar from './components/Navigation/NavBar/NavBar';
 import CSS from './App.module.css';
 
 class App extends Component {
