+++ front_end/src/containers/Auth/Login/Login.js
@@ -3,8 +3,8 @@ import { Link, Redirect } from "react-router-dom";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
 
-import { login, loadUser } from "../store/actions/auth";
-import store from '../store';
+import { login, loadUser } from "../../../store/actions/auth";
+import store from '../../../store';
 
 
 export class Login extends React.Component {
