+++ front_end/src/containers/Auth/Register/Register.js
